move Nature-related classes to Natures.scala
